Replace explicit type declarations with `var`
It looks like you started using `var`, on lines 45..54, but then stopped.
It is generally accepted as quite readable (and in most cases *more* readable) to use var in place of explicit type declaration when an assignment is made. Unless they told you in class to *not* do that, change all the rest of them to use `var`.